add function to script.js
I added a new function called playRound with the parameters defined as
playerSelection and computerSelection. This function was created to play
one round of "Rock, Paper, Scissors" and return a string that states who
won the round or if it was a tie. It also console logs the player and
computers selections for each round.