add function and remove lines in script.js
I added a new function called playAgain. This function is called after
playGame(). It prompts a a confirm dialog box asking the player if they
would like to play again. If the player selects "OK" then it clears the
console and calls playGame() so the game restarts.

Also, I have removed the lines in playGame() that console logged the
score at the end of each round. I found it to be useless and would just
clutter the console.